test: fix ArchivalTest.test_isolate
Test was failing because the topic is created before firewall is closed,
and sometimes topic manifest is uploaded to the S3 bucket before the
firewall is blocked. Changed the assertion to allow topic manifest to be
present in the bucket. Neither segments nor partition manifest should be
present in the bucket.