rpk: bump kadm to fix transactional-id bug, fix doc wording
kadm had a mistaken line duplication which resulted in it being
impossible to list or delete ACLs by transactional ID (the filter code
is the same). This pulls in a patch to fix that.

We also fix some doc wording in ACLs to make a paragraph and help flag
clearer.

This is a recreation of commit a9bb1f357c1d7d2d669d03a710d403277a4ac975
from PR #2982, which did not cherry-pick cleanly due to go.mod and
go.sum dependency changes.