cluster: guesstimate partition leaders on create cmd apply
This way all the nodes have
the same initial leader guess for newly created topics, rather
than the node that serviced the create having a leader while
other nodes do not.

This avoids leadership appearing to flap directly after creation,
if a client sends metadata requests to a different node than they
sent the creation to.

Fixes: https://github.com/vectorizedio/redpanda/issues/2546

Signed-off-by: John Spray <jcs@vectorized.io>