rm_stm: mitigate idempotency violation
Currently Redpanda doesn't do de duplication and returns OOOS error
when it suspects and an out of order record. Modern clients consider
it as a non fatal error, bumps an epoch, resets seq number and retries
the request. As a result in case of a network isolation (client <->
leader) or an leader termination (kill -9) a record may be produced
twice.

Mitigating the problem by returning a generic fatal error.

https://github.com/vectorizedio/redpanda/issues/3039