c/consensus: demote nodes removed from quorum to learners
Fixed issue leading to unavailability that may occur during
reconfiguration in face of failures.

Assumptions:

- raft group with single node
- configuration is described as

```
{current: {voters: [...], learners: [...]}, old: {voters: [...],
learners: [...]}}
```

where each array `[...]` is an array of node ids

Example:
At some point single replica partition was requested to move from one
node to
another, its configuration is going to change.

In Raft in order for this to happen configuration will need to evolve in
a
following steps (assuming we move replica from node 1 to 2)

1. `{current: {voters: [1], learners: []}, old: null}`
2. configuration change requested  from `1` -> `2`
3. `{current: {voters: [], learners: [2]}, old: {voters: [1], learners:
   []}}`
4. after node 2 log is complete it is promoted to be a voter
5. `{current: {voters: [2], learners: []}, old: {voters: [1], learners:
   []}}`
6. after configuration from previous step is committed, leader initiates
   leaving joint consensus mode
7. `{current: {voters: [2], learners: []}, old: null}`
8. after configuration from previous step is committed, old leader steps
   down

In this scenario raft group is vulnerable to availability loss when
current
leader crashes after appending configuration from step 7 or
RPC request is not delivered to node `2` and leadership is lost.

In this scenario last configuration on node `1` is:
```
{current: {voters: [2], learners: []}, old: {voters: [1], learners: []}}
```

while on the node `1`:

```
{current: {voters: [2], learners: []}, old: null}
```

since node `1` appended configuration to its log already, its log is
longer
than on node `2`. This makes all vote requests from `2` to fail.
Node `1` can not be elected as a leader since in its state it is not
longer
a part of a raft group.

In order to solve that unavailability issue we are going to introduce
additional step between 5 & 7. In the additional step we demote all the
voters
from old configuration to learners.

1. `{current: {voters: [1], learners: []}, old: null}`
2. configuration change requested  from `1` -> `2`
3. `{current: {voters: [], learners: [2]}, old: {voters: [1], learners:
   []}}`
4. after node 2 log is complete it is promoted to be a voter
5. `{current: {voters: [2], learners: []}, old: {voters: [1], learners:
   []}}`
6. after configuration from previous step is committed we demote old
   voters to learners
7. `{current: {voters: [2], learners: []}, old: {voters: [], learners:
   [1]}}`
8. after configuration from previous step is committed, leader initiates
   leaving joint consensus mode
9. `{current: {voters: [2], learners: []}, old: null}`
10. after configuration is committed, old leader steps down

In case of a failure described previously, between steps 5 & 7,
raft group will be able to make progress.

configuration on node `1`:

```
{current: {voters: [2], learners: []}, old: {voters: [], learners: [1]}}
```

configuration on node `2`:

```
{current: {voters: [2], learners: []}, old: {voters: [1], learners: []}}
```

in this scenario we must allow learners from old configuration to vote,
this is safe since they must be up to date as they were a voters
previously,
additionally all the checks while casting votes will apply.
In this scenario node `1` will become a leader and continue
to drive the configuration change forward.

If a failure would occur between steps 7&9 nodes configuration would be
following:

configuration on node `1`:

```
{current: {voters: [2], learners: []}, old: null}
```

configuration on node `2`:

```
{current: {voters: [2], learners: []}, old: {voters: [], learners: [1]}}
```

In this scenario node `2` will immediately win election since learner
is not a part of quorum.

Signed-off-by: Michal Maslanka <michal@vectorized.io>