kafka: dedupe last 5 batches
Redpanda didn't support deduplication and used to fail a request when
it noticed that it was out of order. It created an inconvenience for
the users (the only thing they could do in this case is to recreate a
producer) and affected performance characteristics.

Fixing the problem by caching the offsets of the last five replicated
batches. We use five because when idempotence is enabled five is the
max possible value for max.in.flight.requests.per.connection.

Fixes: https://github.com/vectorizedio/redpanda/issues/3038