storage: move index state encoder to compat header
Moves encoding/decoding/checksumming for the to-be-deprecated format
into a compatibility header and out of the main implementation. The
decoder will be used to support the old verison of the file format so we
do not need to rebuilt all indices, while the encoder will be used only
in tests. However, this patch merely re-organizes the code.

Signed-off-by: Noah Watkins <noah@vectorized.io>