k/fetch: return empty records set when max is smaller than start offset
It may happen that last stable offset is equal to log start offset.
In this case when consumer requests a read with a `read_committed`
isolation level it should receive no batches since the log is empty.
Otherwise when records are present in the log we need to read up to
LSO-1  since batch pointed by LSO is the first offset for which
transactions may not be resolved.

Added explicit check to partition read code path to return fast
when requested max offset is smaller than start offset indicating no
data should be returned.

Signed-off-by: Michal Maslanka <michal@vectorized.io>