r/consensus: using cfg manager when there is no delta in snapshot metadata
Recently we added a delta field to raft snapshot metadata. The delta
field is used to recover `offset_translator` state when there is not
enough information in log. When updating from previous redpanda versions
we must support a situation in which delta field may not be present. In
this case we must fallback to previous behavior and use configuration
manager as a source of offset delta.

Signed-off-by: Michal Maslanka <michal@vectorized.io>