cluster: convert members_manager to new join rpc
Only use of old one is now in handling incoming RPCs
from old versions.

This means that new-versioned redpanda will only
be able to join new-versioned clusters.  That
would only impact someone who tried to join
an old version to a newer cluster, or someone
tryin to join an old version to a cluster in
the middle of a rolling upgrade.

Signed-off-by: John Spray <jcs@vectorized.io>