coproc/tests: Fix test_coproc_delete_topic test
- This test was failing in CI due to a race between deletion of
materialized topic and completion of the coproc fiber.

- Its decided to not add a tool to synchronize around this as in real
production workloads a deletion could occur at any time. So this type of
chaos is actually idea for this type of test.

- So the fix is to just apply a different assertion depending on what
async action completed first. If the deletion occured after the fiber
finished processing, assert that the topic doesn't exist, otherwise
assert that all data exists.