k/produce: map raft shutdown error code to kafka timeout error
In order not to wait for all pending replications requests to finish
before shutting down redpanda raft group we simply abort waiting and
exit with `shutting_down` error code. The shutdown may happen either
before or after requests were accepted by the followers. In order to
indicate lack of determinism in this case we return a timeout error,
this indicates that the result of operation is uknown and has to be
verified.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit ebf9e71405197674ff683ddfaee3061a51ec541b)