rm_stm: make abort & commit co-idempotent
rm_stm is a replicated state machine running on top of a Redpanda data
partition so we may break down an execution of a command in a familiar
sequence of steps:

  1) validation
  2) replication
  3) catching up
  4) execution

After replication is over we know than it won't be lost and for some
commands we may decide to return before waiting for step four. But
in this case when we process the next command we can't be sure that
it sees an up to date state.

Caching the replicated offset available after step 2 to use it later
(during an execution of the abort & commit commands) to explicitly
check / wait the previously replicated command is already executed