cluster: Add global shadow indexing toggles
Add two new cluster level configs:
- cloud_storage_enable_remote_read
- cloud_storage_enable_remote_write

The parameters can be used to enable shadow indexing (both archival
uploads and shadow indexing fetch) for all topics. The topic level
configurations (redpanda.remote.read and redpanda.remote.write) act as
an override for the global parameters.

The overrides can be used to enable shadow indexing (or its part) if
shadow indexing is disabled in the config. And also the topic level
overrides can be used to disabled shadow indexing for the topic if
shadow indexing is enabled in the config.