cluster: Test global shadow indexing flags
Add alter-config test that checks that the topic is described correctly
if shadow indexing is enabled globally. Also, the test checks if shadow
indexing can be disabled for a topic if it's enabled globally.

Also, update e2e_shadow_indexing_test to use global settings instead of
per topic configurations. This verifies that shadow indexing can be
enabled globally.