c/controller_backend: fixed finding initial revision during bootstrap
Fixed determining the initial revision of a partition that is performed
during controller_backend bootstrap. Previously it might happen that the
revision was incorrectly determined and the partition replica was
created with incorrect revision which caused `controller_backend` to
stop making progress. In the previous implementation when partition was
first created on a given node with update operation that was preceded by
partition creation `controller_backend` used addition delta to create
partition even tho the addition did not contain the node which creates
partition.
Consider the following set of partition reallocations:

```
[id,id,id] <- partition replica set

add: [1,2,3]
update: [4,2,3]
update  [4,2,1]
```

When bootstrap would be executed on node 4, it should use revision of
[4,2,3] update, while it would use [1,2,3] which would lead to situation
in which one of the partition replicas would be not available.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit 1e46b34bcea37edace3b7f92bb5366b7b01ad1eb)