r/recovery_stm: refresh follower state when snapshot delivery failed
When snapshot is delivered to the follower the request may fail from
different reasons. In this situation leader should refresh follower
state with heartbeat request and retry sending snapshot. Not refreshing
the state may lead to situation in which leader will constantly try to
send a snapshot which can not be accepted by the follower.

Signed-off-by: Michal Maslanka <michal@vectorized.io>