r/recovery_stm: use snapshot index to decide if follower needs a snapshot
When determining if follower requires recovery with a snapshot we used
log start offset. If follower log end was behind the leader log start
offset we deliver a snapshot. Cleanup policy truncates log prefix after
the snapshot has been created (in next garbage collection round in log
manager). This way leader may try to send to the follower data which
should already be removed. Changed the logic in recovery stm to use raft
`_last_included_snapshot` index to determine if follower needs a
snapshot. This way when sending data to follower leader will not use
data from the segment that is about to be deleted but will send follower
a snapshot instead.

Signed-off-by: Michal Maslanka <michal@vectorized.io>