test: handle file removal race in remote checksum
The data_checksum helper in the redpanda service executes the following
command remotely on a node:

   find <path> -type f -exec md5sum '{}' \; -exec stat -c %s '{}' \;

In https://github.com/vectorizedio/redpanda/issues/3258 it was reported
that parsing of this output had the word `No` in the output where an
integer was expected (see data_checksum for more info). One plausible
explanation for this is that a file was deleted between being discovered
by `find` and being passed to an invocation of `md5sum`, resulting in an
output like this:

   >>> "md5sum: {/var/lib/redpanda/data/redpanda/controller/0_0/2-2-v1.log}: No such file or directory".split()
   ['md5sum:', '{/var/lib/redpanda/data/redpanda/controller/0_0/2-2-v1.log}:', 'No', 'such', 'file', 'or', 'directory']

where `No` appeared in the same position as the original `No` that was
unexpected.

This patch changes data_checksum routine to ignore files that appear to
have been removed during the iteration / checksumming process, and logs
them at debug level for debugging if necessary.

Fixes: #3258

Signed-off-by: Noah Watkins <noah@vectorized.io>