archival: Fix data corruption issue
The bug manifested as a truncated segment in S3. This happens only if
the cloud_storage_segment_max_upload_interval_sec option is enabled.

This is what causes the bug.  The code reads the size of the segment at the
begining and uses it to set the initial content-length value. Then it
performs an asynchronous operation. During this asynchronous operation
the size of the segment might change but the content-length value will
stay the same.

After that we're getting the segment size once again and here we can have
a larger value. This value is evenutally used to compute the size of the tail
region of the segment which shouldn't be uploaded (file size minus the
position where the upload should stop). Because the code uses the new
file size here it subtracts larger value which results in truncation
of the segment.

The fix is to just read the size of the segment once.

(cherry picked from commit 3a252512a8b7a0fbadfde7ff33883e67528fd7e0)