archival: Use I/O priority for partial uploads
When the cloud_storage_segment_max_upload_interval_sec option is used
archival subsystem tries to read from segment to find the right
locations of the begining and the end of the segment. This reads didn't
use the right io_priority_class that the rest of the archival uses. This
commit fixes this by propagating the right io_priority_class.

(cherry picked from commit 52853530823836c18baddfebe1fb7af8be990b07)