cloud_storage: Stop partition_record_batch_reader_impl
The partition_record_batch_reader_impl component is not stopping when
the underlying remote_partition is stopped. This manifested in the
following situation during failure. The reader stuck in an infinite loop
first. Then the remote_partition was stopped, but the infinite loop
didn't. It continued to consume CPU even when the entire topic was
deleted.

Thic commit fixes this by checking the abort_source inside the
remote_partition. Fixes #3305