cloud_storage: don't use segment paths as manifest keys
Previously segment paths were used in the manifest for the case when
segments come from a different ntp revision than the current one
(these segments could appear in the manifest for example after
recovery). This lead to a lot of code that had to handle both cases for
manifest keys (short names and full paths) and parsed these paths and
assembled them back. By storing ntp revision id in segment_meta instead,
we can eliminate all this code. Segment paths become opaque, as we don't
have to parse them anywhere - we only generate them when we need to talk
to cloud storage and that's it.