cloud_storage: Store file obj inside remote_segment
Previously the cache interface returned the input_stream and size of
the file. Because of that the file object was lost and couldn't be
properly closed.

In order to fix this the cache interface was changed. Method 'get' now
returns cached_item that has ss::file instance and not the
ss::input_stream<char>. Because of that some parameters were removed
from the method 'get' (like pos, io-priority, etc). These parameters
were used to initialize ss::input_stream<char>.

The remote_segment was changed to accomodate new interface and to
actually store the file object. The background hydration loop has a fast
path for the situation when remote_segment already has a file instance.
If it doesn't have it it hydrates the segment and then uses method 'get'
of the cache to get the file object.

The ss::input_stream<char> instance is now created inside the
remote_segment instance.