tests: use majority to determine if node is removed
In redpanda the `v1/brokers` API return a list of brokers that is stored
in replying node metadata. The view may be stale and do not account all
the cluster membership changes.
When node is removed from the cluster it is not longer present in the
list of nodes returned from `v1/brokers` API, however a node may receive
an answer from the node that metadata are out of date especially when
the node was experiencing failure

The change introduced in this commit may be reverted after we will
introduce an API that will allow user to perform linearized read of
cluster state.

Signed-off-by: Michal Maslanka <michal@vectorized.io>