net/connection: introduced per connection gate
Introduced a per connection gate to control the connection life cycle.
Without the gate it was possible that connection was shutdown during
processing an RPC request. This way when the request was finished its
result wasn't propagated to requester. Introduction of per connection
gate will allow us to wait for all requests pending for given connection
before shutting it down.

Signed-off-by: Michal Maslanka <michal@vectorized.io>