cloud_storage: refine logging & exceptions on remote_segment
Including the path in log messages makes it much easier
to chase down issues with the state for particular
files (such as references to deleted files).

Adds an assert in ~remote_segment that the data file
was propertly closed, to increase confidence that
the FD is not leaked in this class.

Additionally use an `optional` for `_data_file` to make
checks for its set-ness more explicit and clear.

Signed-off-by: John Spray <jcs@vectorized.io>