cloud_storage: fix file descriptor leak
The garbage collection of remote_segment objects
was conditional on the reference count in the
segment's sharded_ptr being 1 (i.e. nothing else
is using it).

This is not true if there are entries in the `readers`
attribute of materialized_segment_state, because
readers hold a reference to the segment.

Deal with this by proactively evicting readers
from any materialized_segment_state objects
whose atime makes them elegible for removal.

Fixes: https://github.com/vectorizedio/redpanda/issues/3334

Signed-off-by: John Spray <jcs@vectorized.io>