c/dissemination: move arming dispatch timer to service start method
Dissemination dispatch timer should not be started in dissemination
service constructor as some of the dissemination service dependencies
may not be ready when it is constructed.

The `metadata_dissemination_service` is constructed and started after the
`controller` so it should be fine to use controller APIs in the
service. However we split services construction from starting them.
Starting a timer in constructor may lead to situation in which
controller would be accessed before it was started.

Signed-off-by: Michal Maslanka <michal@vectorized.io>