bump kadm && task tidy
This bumps kadm to a version that uses per-partition errors if
OffsetCommitResponse is missing some topics/partitions in the response
that were in the request.

Notably, this was seen when seeking against redpanda: redpanda does not
today either match the request ordering, or does not respond to
everything all of the time. This is tracked in #3403.