c/metadata_cache: fixed returning alive brokers
Fixed issue causing `metadata_cache::all_alive_brokers` to always return
all brokers even if of the brokers in the cluster are down.

Expected behavior of `all_alive_brokers` is to return only those brokers
which are alive or in case in which there are no alive brokers list of
all brokers in the cluster. In scenarios when brokers are marked as not
alive by health monitor they should be filtered out from the returned
brokers list.

The issue was caused by incorrect update of `brokers_with_health` variable.
When any of the broker was down the logic automatically returned full
brokers list.

Signed-off-by: Michal Maslanka <michal@vectorized.io>