archival: Make upload loop more resilent
Handle all exceptions in the upload loop. If one of the archivers will
throw an exception which is not the expected kind just add the message
to the log. On the next iteration we will proceed with different set of
archivers so broken archiver won't cause unavailability for all
partitions.

This change is supposed to make transient errors less severe. For
instance, it is possible for the log to get truncated while the archiver
is trying to upload it. It's safe to assume that this archiver can retry
after a while. The archiver would be pushed to the back of the upload
queue so it won't be retried right away.

(cherry picked from commit 3c7fcda3a55b2e08ff35a671e3c4101f285531ff)