c/controller_backend: move offset translator state when moving partition
When partition is moved x-cores all the persistent state stored in core
local kvstore should be moved together with the `raft::consensus`
instance. Offset translator state is recoverable so moving its state is
an optimization. However, moving partition x-cores should not be
expensive as it does not involve data movement. In order to avoid
reading the whole log while recovering the offset translator state we
move its state from `kvstore` on source shard to the destination shard
`kvstore`

Signed-off-by: Michal Maslanka <michal@vectorized.io>