cluster: pass latest version back from config_frontend::patch
Since remote execution of config updates was added,
the node handling the admin API call can no longer
use its local version as the new latest version
in its response to clients.

Augment the return value throughout to include an error_code
and a config_version, instead of just the error code.  Then
use that in the return to the admin API client.

Signed-off-by: John Spray <jcs@vectorized.io>