cloud_storage: suppress warning on successfull topic creation
When topic is created with Shadow Indexing enabled, ntp_archiver starts
and tries to download a manifest from S3 before the manifest is
uploaded. That's a normal situation and should not create warnings.
Remote.cc now only logs errors about timeouts and retries, a decision
how to handle not_found happens on a higher level where we have a
context about situation.