storage: Add initial revision to ntp_config
Initial revision id is a number that indicates the first revision that
NTP had when the topic was just created. The value is used by the shadow
indexing subsystem to provide stable naming schema. Segment names
contains revision id. Also the revision id can change when the partition
is moved between the nodes. To avoid this the initial revision id is
introduced to the NTP config. It will be a part of the segment path in
the S3 bucket to avoid scattering of the partition data.