s/disk_log_impl: serialize start_offset updates
See https://github.com/vectorizedio/redpanda/issues/3377. Critical
invariant for _start_offset is that it never decreases.

Previously this invariant could be violated in a couple of ways.
* Because _start_offset was updated only at the end of _truncate_prefix
function, if all segments were deleted, external observer could
observe _start_offset going backwards to the previous value after all
segments have been deleted, but before it was updated in memory.
* Because kvstore updates were not serialized, we could never be sure
that we are not overwriting the bigger _start_offset value that gets
written as a part of concurrent prefix_truncate.

To mitigate that we serialize the kvstore updates with a mutex and
update in-memory _start_offset at the beginning of prefix_truncate.

(cherry picked from commit 038d6c82712983cd4fdf19f583a652d24693de74)