test: remove generic kafka client interface
This was an experiment for creating a fully generic client interface
that we can swap out implementations and where we could grab different
versions of the underlying client. This didn't fully workout and plan to
take a more practical appraoch involving building a common default
client and then supporting specialized clients for specific testing.

Signed-off-by: Noah Watkins <noah@vectorized.io>