test: add default kafka client
The intention of this is to collect the vast majority of kafka client
use cases and expose the client through the test base class for tests to
make use of (e.g. without creating clients explicitly). Tests that want
to use specific clients can continue to do so.

Initially this default client is collecting use cases factored out of
the redpanda service and an assortment of tests, and as such is using
more than one underlying client. The final step will be rework this
client and converge around a single underlying client such as
rpk/franz-go.

Signed-off-by: Noah Watkins <noah@vectorized.io>