test: fix type conversion error
Type str cannot be converted into ReadableBuffer. Instead, encode the
string into bytes. Encoding of ascii is used for strictness, which
should be a realistic restriction for namespace, topic, partition. It
will be easier to catch a violation and generalize than to assume ascii
but let utf-8 to pass through.

Signed-off-by: Noah Watkins <noah@vectorized.io>