tools: expose the test's unique folder in dash
The first table within the coverage dashboard summarizes coverage totals
for each ducktape tests. In the table, the name of the test should link
to the sub-dir with code coverage files such as index.html and
coverage.json. Those files provide more details for coverage results.

Exposing the sub-dir makes the files downloadable instead of listing
each file for each test in the buildkite artifacts tab.