k8s: use statefulset revision for pod update
Currently we're using patchmaker to compare pods to figure out if we
need to delete them to continue rolling update. Patchmaker is pretty
complex thing to get right to not end up in situation where changes are
constantly detected or not detected at all. So we rather fall back to
implementing the exact same comparison the sts is using to implement
rolling update to implement our update strategy.