test: remove optional return type
rpk.describe_topic_configs had been returning Optional[dict] in the case
that the target topic didn't exist. However, this means that the python
type checker will report errors for all users of this that don't
explicitly check for None return type. All current users of this method
assume that None is not returned. This change eliminates 26 type errors.

Signed-off-by: Noah Watkins <noah@vectorized.io>