tests: make metrics_reporter_test more reliable
...and extend it to check that cluster parameters
survive a restart.

This test was previously starting redpanda twice: once
implicitly in setUp, then again implicitly in the test.

Depending on the speed of the test node, this could
result in a failure where on the initial startup, a
report is sent with a different cluster uuid+timestamp
than on subsequent startup.  Presumably this corresponds
to a config batch that is generated but never committed
because the nodes are restart so soon after when
the test starts.

Fix this by only starting redpanda once, and add an explicit
test that on restart (/after/ the system has been up long enough
to send several reports and presumably have committed its original
config message), these metric fields do not change.

Fixes: https://github.com/vectorizedio/redpanda/issues/3514
(cherry picked from commit fcac9f112811970146bcd3f615b379da39d24c05)