storage/snapshotable_stm: make snapshot in background
The `storage::snapshotable_stm` interfaces clearly states that the
`make_snapshot` function is called to give an stm implementation a hint
on when to create a snapshot. There is requirement to wait for a
snapshot to be taken before returning from `disk_log_impl::new_segment`
function.

Now `snapshotable_stm::make_snapshot_in_background` returns `void` and it is up
to the implementation on what to do when the function is called.
Peristed stm implementation is dispatching taking a snapshot in
background fiber.

This change fixes an issue described in #3582, as the deadlock was
caused by the fiber applying batches and appending batches to wait for
stm lock. Currently as the snapshot is taken in the background the
deadlock is impossible to occur.

Fixes: #3582:

Signed-off-by: Michal Maslanka <michal@vectorized.io>