ssx: add ssx::spawn_with_gate, ssx::background
ssx::background is a more explicit alternative to
ignoring futures by casting to (void), similar to
std::ignore.

Spawn is for places where we currently do (void)ss::with_gate,
which is error-prone because it can result in unhandled
exceptions during shutdown.  Any place we're detaching a
background fiber with a gate, we should be handling
the shutdown exceptions as a no-op so that the background
fiber terminates quietly when process is shutting down.

Fixes https://github.com/vectorizedio/redpanda/issues/2911