tests: include example in BadLogLines exception string
This previously just said which nodes had bad lines.  It's
common that a test failure comes from just a single line,
so to help in that case we can include the first bad line
as an example.  This will show up at the bottom of a test
log & save effort scrolling back to the offending log line.