admin: improve leader redirect heuristic
Previously we guessed that the leader would
have an admin API listener on their internal RPC
address, and that it would be on the same port as
the local node's first admin listener.

This worked fine when using internal pod IPs,
but falls down under other circumstances.  This
was okay, because we already fell down (with a 400)
on non-leader requests.

However, it was more problematic when updating rpk,
because rpk doesn't know whether it's sending
a redirect-able request or not.

We can make the redirection smarter, to handle the
typical case where both admin and kafka have
two listeners, one on an internal name and one
on an external name:
- Compare the incoming request's Host header to
  the local node's kafka listeners to find which
  kafka-equivalent address the client was using (i.e.
  is it an internal or external request?)
- Pick the same index in the peer's array of kafka
  advertised addresses.  On simple uniform internal+external
  configs like in our cloud, that will reliably select
  the peer's appropriate internal/external DNS name
  for us to redirect the admin API request to.