kafka/client: Handle topic_error in mitigate_error
The `topic_error` can be emitted from the `topic_cache` when producing.

The fix is to retry with backoff to update metadata, on the assumption
that the topic will become ready.

The missing behaviour was detected in
`rptest.tests.pandaproxy_test.PandaProxyTest.test_produce_topic`
as `BadLogLines`.

Also fixes a missing fmt placeholder for `partition_error`.

Signed-off-by: Ben Pope <ben@vectorized.io>